pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "eval-crypt"
version = "0.1.0"
description = "A tool to encrypt sensitive AI safety evaluation files in Git repositories"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click",
    "pycryptodome",
    "pre-commit",
]

[project.scripts]
eval-crypt = "eval_crypt.cli:main" 

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-click",
    "pre-commit",
    "ruff",
    "mypy"
]

[tool.setuptools]
packages = [
    "eval_crypt"
]